Renaming / refactoring server
This moves server to the right location and just leaves connecting
locally left to deal with
